msgpack: move field related code to field.go
